Actually enable spotbugs (#75)
Switch to HTML output.

Baseline against current status after fixing a few bugs.

Archive html.

While there: bump dependencies.

Add junit to artifacts through built-in and touch files to avoid time limits.

Remove Serializable from the MPIInterface and other minor fixes.

+++ src/org/ogolem/rmi/GenericGlobOptJob.java
@@ -260,7 +260,8 @@ public class GenericGlobOptJob<E, T extends Optimizable<E>> implements Job<T> {
       return new ArrayList<>();
     }
 
-    final int noTasks = Math.min(maxTasks, (int) Math.ceil(this.poolSize / noProxies));
+    final int noTasks =
+        Math.min(maxTasks, (int) Math.ceil((double) this.poolSize / (double) noProxies));
 
     if (DEBUG) {
       System.out.println(
